+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [unreleased]
 
+- Switched QR code scanning to scan full page and report orientation
+- Added second attempt for QR code scanning with higher default resolution, `try_rotate=True`, and `try_downscale=True`
+
 ## [0.3.0]
 
 - Adopt uv for project structure